requirements.txt
pymupdf
img2pdf
Pillow
requests
azure-functions
